Fix Storybook production crash from server-only barrel imports
app/utils/index.ts mixed pure client utilities (colorSwitcher, etc.)
with server-only code (node:fs, the "use server" Sanity action). Importing
colorSwitcher dragged the entire server graph into the client bundle,
causing a "before initialization" error in the minified production build
(e.g. the ViewLink story) while working fine in dev.

Move server-only functions (fileToString, generateBlogMeta) to
app/utils/server.ts, keep index.ts pure, and update server-side importers.
