+++ .gitignore
@@ -38,3 +38,4 @@ next-env.d.ts
 
 # storybook
 debug-storybook.log
+storybook-static/
